refactor utils.mongo
