+++ commons/src/main/java/uk/ac/ebi/biosamples/utils/mongo/SampleReadService.java
@@ -8,7 +8,7 @@
 * CONDITIONS OF ANY KIND, either express or implied. See the License for the
 * specific language governing permissions and limitations under the License.
 */
-package uk.ac.ebi.biosamples.service;
+package uk.ac.ebi.biosamples.utils.mongo;
 
 import java.util.List;
 import java.util.Optional;
