+++ pipelines/curami/pom.xml
@@ -34,7 +34,7 @@
 		</dependency>
 		<dependency>
 			<groupId>uk.ac.ebi.biosamples</groupId>
-			<artifactId>utils-mongo</artifactId>
+			<artifactId>commons</artifactId>
 			<version>5.2.3-SNAPSHOT</version>
 		</dependency>
 		<dependency>
