+++ pipelines/reindex/pom.xml
@@ -15,7 +15,7 @@
 	<dependencies>
 		<dependency>
 			<groupId>uk.ac.ebi.biosamples</groupId>
-			<artifactId>utils-mongo</artifactId>
+			<artifactId>commons</artifactId>
 			<version>5.2.3-SNAPSHOT</version>
 		</dependency>
 		<dependency>
