+++ pipelines/reindex/src/main/java/uk/ac/ebi/biosamples/ReindexRunner.java
@@ -28,7 +28,7 @@ import org.springframework.data.util.CloseableIterator;
 import org.springframework.stereotype.Component;
 import uk.ac.ebi.biosamples.model.Sample;
 import uk.ac.ebi.biosamples.mongo.model.MongoSample;
-import uk.ac.ebi.biosamples.service.SampleReadService;
+import uk.ac.ebi.biosamples.utils.mongo.SampleReadService;
 import uk.ac.ebi.biosamples.utils.ThreadUtils;
 
 /**
