+++ pipelines/reindex/src/test/java/uk/ac/ebi/biosamples/ReindexRunnerTest.java
@@ -25,7 +25,7 @@ import org.springframework.data.mongodb.core.query.Query;
 import org.springframework.data.util.CloseableIterator;
 import uk.ac.ebi.biosamples.model.Sample;
 import uk.ac.ebi.biosamples.mongo.model.MongoSample;
-import uk.ac.ebi.biosamples.service.SampleReadService;
+import uk.ac.ebi.biosamples.utils.mongo.SampleReadService;
 
 @RunWith(MockitoJUnitRunner.class)
 public class ReindexRunnerTest {
