+++ pipelines/zooma/src/main/java/uk/ac/ebi/biosamples/zooma/ZoomaApplicationRunner.java
@@ -32,7 +32,7 @@ import uk.ac.ebi.biosamples.client.BioSamplesClient;
 import uk.ac.ebi.biosamples.model.PipelineAnalytics;
 import uk.ac.ebi.biosamples.model.Sample;
 import uk.ac.ebi.biosamples.model.filter.Filter;
-import uk.ac.ebi.biosamples.service.AnalyticsService;
+import uk.ac.ebi.biosamples.utils.mongo.AnalyticsService;
 import uk.ac.ebi.biosamples.service.CurationApplicationService;
 import uk.ac.ebi.biosamples.utils.AdaptiveThreadPoolExecutor;
 import uk.ac.ebi.biosamples.utils.ArgUtils;
