--- utils/mongo/pom.xml
@@ -1,33 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-
-	<artifactId>utils-mongo</artifactId>
-	<packaging>jar</packaging>
-
-	<parent>
-		<groupId>uk.ac.ebi.biosamples</groupId>
-		<artifactId>biosamples</artifactId>
-		<version>5.2.3-SNAPSHOT</version>
-		<relativePath>../../</relativePath>
-	</parent>
-
-	<dependencies>
-		<dependency>
-			<groupId>uk.ac.ebi.biosamples</groupId>
-			<artifactId>models-mongo</artifactId>
-			<version>5.2.3-SNAPSHOT</version>
-		</dependency>
-		<dependency>
-			<groupId>uk.ac.ebi.biosamples</groupId>
-			<artifactId>properties</artifactId>
-			<version>5.2.3-SNAPSHOT</version>
-		</dependency>
-		<dependency>
-			<groupId>uk.ac.ebi.biosamples</groupId>
-			<artifactId>utils-thread</artifactId>
-			<version>5.2.3-SNAPSHOT</version>
-		</dependency>
-	</dependencies>
-
-</project>
