+++ utils/webapp/src/main/java/uk/ac/ebi/biosamples/service/MessagingService.java
@@ -28,6 +28,7 @@ import uk.ac.ebi.biosamples.model.Sample;
 import uk.ac.ebi.biosamples.model.StaticViewWrapper;
 import uk.ac.ebi.biosamples.mongo.repo.MongoSampleRepository;
 import uk.ac.ebi.biosamples.mongo.service.SampleToMongoSampleConverter;
+import uk.ac.ebi.biosamples.utils.mongo.SampleReadService;
 
 @Service
 public class MessagingService {
