+++ webapps/core/src/main/java/uk/ac/ebi/biosamples/service/StatService.java
@@ -22,6 +22,7 @@ import uk.ac.ebi.biosamples.model.filter.Filter;
 import uk.ac.ebi.biosamples.mongo.model.MongoAnalytics;
 import uk.ac.ebi.biosamples.solr.service.SolrFacetService;
 import uk.ac.ebi.biosamples.solr.service.SolrFieldService;
+import uk.ac.ebi.biosamples.utils.mongo.AnalyticsService;
 
 @Service
 public class StatService {
